- Refactor old ds dir away - Rename files to be more consistent
